+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/panels/RowBuilder.kt
@@ -11,16 +11,23 @@ import ru.yandex.money.tools.grafana.dsl.generators.PanelLayoutGenerator
 class RowBuilder(
     title: String,
     repeat: String?,
-    override val panelLayoutGenerator: PanelLayoutGenerator
+    override val panelLayoutGenerator: PanelLayoutGenerator,
+    private val collapsed: Boolean = false
 ) : PanelContainerBuilder {
 
-    override val panels = mutableListOf<Panel>(
-            Row(BasePanel(
-                id = panelLayoutGenerator.nextId(),
-                title = title,
-                position = panelLayoutGenerator.nextPosition(24, 1)
-            ), repeat)
-    )
+    private val row: Row = Row(BasePanel(
+        id = panelLayoutGenerator.nextId(),
+        title = title,
+        position = panelLayoutGenerator.nextPosition(24, 1)
+    ), repeat, collapsed)
 
-    internal fun createRow() = panels.toList()
+    override val panels = mutableListOf<Panel>(row)
+
+    internal fun createRow() = if (!collapsed || panels.size == 1) {
+        panels.toList()
+    } else {
+        val row: Row = row
+        row.panels = Panels(panels.drop(1).toList())
+        listOf(row)
+    }
 }